refactor(command): organize user commands in subdir
