+++ compiler_admin/commands/user/delete.py
