+++ compiler_admin/commands/user/reset_password.py
@@ -1,7 +1,7 @@
 from argparse import Namespace
 
 from compiler_admin import RESULT_SUCCESS, RESULT_FAILURE
-from compiler_admin.commands.signout import signout
+from compiler_admin.commands.user.signout import signout
 from compiler_admin.services.google import USER_HELLO, CallGAMCommand, user_account_name, user_exists
