+++ tests/commands/user/test_reset_password.py
@@ -2,7 +2,7 @@ from argparse import Namespace
 import pytest
 
 from compiler_admin import RESULT_FAILURE, RESULT_SUCCESS
-from compiler_admin.commands.reset_password import reset_password, __name__ as MODULE
+from compiler_admin.commands.user.reset_password import reset_password, __name__ as MODULE
 from compiler_admin.services.google import USER_HELLO
